requirements.txt
python>=3.10
pandas>=1.3.5
numpy>=1.21.0
scikit-learn>=1.0.2
xgboost>=1.6.2
shap>=0.41.0
wandb>=0.15.0
python-dotenv>=0.21.0
jupyterlab>=3.0.0
ipykernel>=6.0.0
anthropic>=0.3.0
pydantic>=1.10.7
together==1.5.30
google-cloud-storage>=2.10.0
google-genai>=0.2.0
pytest>=7.0.0